Move Sanitizer to Entry/ folder
